Drop timeout on sync.
When out of sync all bytes must be drained at once in order to be able to reach sync again.
This strategy works because the radio will repeatedly send the bootmessage.
We do not need to save bytes and try to sync up based on bytes, just drop all received bytes and wait for the radio to send a new bootmessage.

If we wait 10 seconds (or any time longer then the boot message repeat time) we will read garbage from the previous package, discard those and the rest of the message and we will never reach sync.